Realised enormous refactoring of the Discord bot.

+++ ACBot.Discord/Command/Utilities/ConversionUtilities.cs
@@ -3,7 +3,12 @@
 public static class ConversionUtilities
 {
 
-    public static DayOfWeek? IntegerToDay(long day)
+    /// <summary>
+    /// Convert the given input into a DayOfWeek enum value by its index, if valid.
+    /// </summary>
+    /// <param name="day">The day index for the day of the week.</param>
+    /// <returns>The DayOfWeek enum value if the day is a valid index, null otherwise.</returns>
+    public static DayOfWeek? IntegerToDay(int day)
     {
         if (day is < 0 or > 6)
             return null;
